modified the operator new unsigned variables to unsigned long variables to satisfy the stricter compiler requirements
git-svn-id: svn://svn.code.sf.net/p/simias/code/branches/iFolder_3_2@6532 1a4d904c-e006-0410-be0b-b506c45ff692

+++ simias/external/flaim/include/flaim.h
@@ -1244,18 +1244,18 @@ Desc: This include file contains the structure definitions and prototypes
 		}
 
 		void * operator new(
-			unsigned			uiSize);
+			unsigned long			uiSize);
 
 		void * operator new(
-			unsigned			uiSize,
+			unsigned long			uiSize,
 			char *			pszFile,
 			int				iLine);
 
 		void * operator new[](
-			unsigned			uiSize);
+			unsigned long			uiSize);
 
 		void * operator new[](
-			unsigned			uiSize,
+			unsigned long			uiSize,
 			char *			pszFile,
 			int				iLine);
 
@@ -4250,18 +4250,18 @@ typedef struct
 		}
 
 		void * operator new(
-			unsigned			uiSize);
+			unsigned long			uiSize);
 
 		void * operator new(
-			unsigned			uiSize,
+			unsigned long			uiSize,
 			char *			pszFile,
 			int				iLine);
 
 		void * operator new[](
-			unsigned			uiSize);
+			unsigned long			uiSize);
 
 		void * operator new[](
-			unsigned			uiSize,
+			unsigned long			uiSize,
 			char *			pszFile,
 			int				iLine);
